k256: extract `arithmetic::mul::glv` submodule
Extracts a submodule containing the implementation details of how we
leverage the GLV endomorphism
